extract DevToolsPanel subtree from App.jsx
- DevToolsPanel.jsx: header actions + section composition
- One component per config section: BasicSettings, Progression,
  HealthMetrics, Primitives (accordion), Mechanics, Scoring
- Section JSX moved verbatim; identifiers become same-named props so
  the diff is mechanical. State stays lifted in App so dev-tool edits
  survive switching back to the dashboard view
- Removed dead updateBudgetConfig (never referenced)
- App.jsx: 1,942 -> 972 lines

Closes #30
