+++ client/src/components/devtools/BasicSettingsSection.jsx
@@ -1,0 +1,84 @@
+import React from 'react';
+
+export function BasicSettingsSection({ editConfig, setEditConfig, updateEditConfig }) {
+  return (
+    <>
+          <h3 className="section-divider">Basic Settings</h3>
+          <div className="dev-tools-grid three-col">
+            <div className="config-section">
+              <h3>Starting Values</h3>
+              <div className="config-field">
+                <label>Treasury (£)</label>
+                <input type="number" value={editConfig.startingTreasury} onChange={(e) => updateEditConfig('startingTreasury', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Residents</label>
+                <input type="number" value={editConfig.startingResidents} onChange={(e) => updateEditConfig('startingResidents', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Builds / Week</label>
+                <input type="number" min="1" value={editConfig.buildsPerWeek ?? 1} onChange={(e) => updateEditConfig('buildsPerWeek', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Policy Changes / Week</label>
+                <input type="number" min="1" value={editConfig.policyChangesPerWeek ?? 1} onChange={(e) => updateEditConfig('policyChangesPerWeek', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Research Actions / Week</label>
+                <input type="number" min="1" value={editConfig.researchActionsPerWeek ?? 1} onChange={(e) => updateEditConfig('researchActionsPerWeek', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Ground Rent (£/wk)</label>
+                <input type="number" value={editConfig.groundRentBase} onChange={(e) => updateEditConfig('groundRentBase', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Utilities (£/wk)</label>
+                <input type="number" value={editConfig.utilitiesBase} onChange={(e) => updateEditConfig('utilitiesBase', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Game Over (£)</label>
+                <input type="number" value={editConfig.gameOverLimit} onChange={(e) => updateEditConfig('gameOverLimit', e.target.value)} />
+              </div>
+            </div>
+
+            <div className="config-section">
+              <h3>Rent & Churn</h3>
+              <div className="config-field">
+                <label>Min (£)</label>
+                <input type="number" value={editConfig.rentMin} onChange={(e) => updateEditConfig('rentMin', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Max (£)</label>
+                <input type="number" value={editConfig.rentMax} onChange={(e) => updateEditConfig('rentMax', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Default (£)</label>
+                <input type="number" value={editConfig.defaultRent} onChange={(e) => updateEditConfig('defaultRent', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Base Churn</label>
+                <input type="number" step="0.01" value={editConfig.baseChurnRate} onChange={(e) => updateEditConfig('baseChurnRate', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label title="How much each £ of rent adds to churn rate. At 0.0003, charging £100 adds 3% to weekly churn.">Rent Impact</label>
+                <input type="number" step="0.0001" value={editConfig.churnRentMultiplier} onChange={(e) => updateEditConfig('churnRentMultiplier', e.target.value)} />
+              </div>
+            </div>
+
+            <div className="config-section">
+              <h3>Timing</h3>
+              <div className="config-field">
+                <label>Tick (ms)</label>
+                <input type="number" value={editConfig.tickSpeed} onChange={(e) => updateEditConfig('tickSpeed', e.target.value)} />
+              </div>
+              <div className="config-field">
+                <label>Base Recruit Slots</label>
+                <input type="number" value={editConfig?.health?.baseRecruitSlots ?? 1} 
+                  onChange={(e) => setEditConfig({...editConfig, health: {...editConfig.health, baseRecruitSlots: parseInt(e.target.value)}})} />
+              </div>
+            </div>
+
+          </div>
+    </>
+  );
+}
